Add CGPoint+
